New Optional Param for naming attachment
René,

This PR brings a new, optional parameter to the requestForCreatingImageAttachment for proscribing a filename, rather than autogenerating one.

This also brings a change to the uploading extension that exercises this new parameter to give a static filename to the MapSnapshot that’s taken and uploaded.

This allows me to filter out the map snapshot in the image carousel used inside Salesforce.